+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This document contains only major functionality changes in certain versions. Min
 ## 4.0
 
 - SDK was separated into 2 packages: `@ringcentral/sdk` and `@ringcentral/subscriptions`
+- `appKey` and `appSecret` were replaced with `clientId` and `clientSecret`
 
 ## 3.0